Restructured file layout. Separated some structs into their own files
