+++ Ranitas/Insects/FlyRenderer.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using Ranitas.Core;
 using Ranitas.Core.Render;
