+++ Ranitas/Sim/ComponentSystems/Components/ControlledEntity.cs
@@ -1,0 +1,14 @@
+﻿using Ranitas.Core.ECS;
+
+namespace Ranitas.Sim
+{
+    public struct ControlledEntity
+    {
+        public ControlledEntity(Entity controlledEntity)
+        {
+            Entity = controlledEntity;
+        }
+
+        public Entity Entity;
+    }
+}
