+++ Ranitas/Sim/ComponentSystems/Components/Player.cs
@@ -1,0 +1,12 @@
+﻿namespace Ranitas.Sim
+{
+    public struct Player
+    {
+        public Player(int index)
+        {
+            Index = index;
+        }
+
+        public int Index;
+    }
+}
